Move error handling code to runtime/error module
