+++ exonum-java-binding/core/rust/integration_tests/tests/runtime_bootstrap.rs
@@ -17,7 +17,7 @@
 extern crate integration_tests;
 extern crate java_bindings;
 
-use integration_tests::vm::{tests_classpath, java_library_path, log4j_path};
+use integration_tests::vm::{java_library_path, log4j_path, tests_classpath};
 use java_bindings::{
     create_java_vm, create_service_runtime, Executor, InternalConfig, JvmConfig, RuntimeConfig,
 };
